Optimize actions
It's not parallel, but requesting actions in groups of 100 is already
pretty good. We spend most of the time waiting for Icinga, anyway: If
one single filter matches 10000 items, we have to wait for Icinga to
process them all. In other words, HTTP overhead does not waste a lot of
time, unless you issue a single HTTP request for each item.